+++ driver-core/src/test/java/com/datastax/driver/core/TokenRangeTest.java
@@ -1,13 +1,10 @@
 package com.datastax.driver.core;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import org.testng.annotations.Test;
 
 import static org.testng.Assert.fail;
-
-import com.datastax.driver.core.Token.M3PToken;
 
 import static com.datastax.driver.core.Assertions.assertThat;
 
@@ -77,6 +74,25 @@ public class TokenRangeTest {
     }
 
     @Test(groups = "unit")
+    public void should_compute_intersection() {
+        assertThat(tokenRange(3, 9).intersectWith(tokenRange(2, 4)))
+            .isEqualTo(tokenRange(3, 4));
+        assertThat(tokenRange(3, 9).intersectWith(tokenRange(3, 5)))
+            .isEqualTo(tokenRange(3, 5));
+        assertThat(tokenRange(3, 9).intersectWith(tokenRange(4, 6)))
+            .isEqualTo(tokenRange(4, 6));
+        assertThat(tokenRange(3, 9).intersectWith(tokenRange(7, 9)))
+            .isEqualTo(tokenRange(7, 9));
+        assertThat(tokenRange(3, 9).intersectWith(tokenRange(8, 10)))
+            .isEqualTo(tokenRange(8, 9));
+    }
+
+    @Test(groups = "unit", expectedExceptions = IllegalArgumentException.class)
+    public void should_fail_to_compute_intersection_when_ranges_dont_intersect() {
+        tokenRange(1, 2).intersectWith(tokenRange(2, 3));
+    }
+
+    @Test(groups = "unit")
     public void should_merge_with_other_range() {
         assertThat(tokenRange(3, 9).mergeWith(tokenRange(2, 3))).isEqualTo(tokenRange(2, 9));
         assertThat(tokenRange(3, 9).mergeWith(tokenRange(2, 4))).isEqualTo(tokenRange(2, 9));
@@ -128,12 +144,12 @@ public class TokenRangeTest {
 
     @Test(groups = "unit", expectedExceptions = IllegalArgumentException.class)
     public void should_not_merge_with_nonadjacent_and_disjoint_ranges() {
-        tokenRange(0,5).mergeWith(tokenRange(7,14));
+        tokenRange(0, 5).mergeWith(tokenRange(7, 14));
     }
 
     @Test(groups = "unit")
     public void should_return_non_empty_range_if_other_range_is_empty() {
-        assertThat(tokenRange(1,5).mergeWith(tokenRange(5,5))).isEqualTo(tokenRange(1,5));
+        assertThat(tokenRange(1, 5).mergeWith(tokenRange(5, 5))).isEqualTo(tokenRange(1, 5));
     }
 
     @Test(groups = "unit")
@@ -155,14 +171,13 @@ public class TokenRangeTest {
         assertThat(splits).containsExactly(tokenRange(3, 5), tokenRange(5, 7), tokenRange(7, 9));
     }
 
-
     @Test(groups = "unit")
     public void should_throw_error_with_less_than_1_splits() {
-        for(int i = -255; i < 1; i++) {
+        for (int i = -255; i < 1; i++) {
             try {
                 tokenRange(0, 1).splitEvenly(i);
                 fail("Expected error when providing " + i + " splits.");
-            } catch(IllegalArgumentException e) {
+            } catch (IllegalArgumentException e) {
                 // expected.
             }
         }
@@ -175,15 +190,15 @@ public class TokenRangeTest {
 
     @Test(groups = "unit")
     public void should_create_empty_token_ranges_if_too_many_splits() {
-        TokenRange range = tokenRange(0,10);
+        TokenRange range = tokenRange(0, 10);
 
         List<TokenRange> ranges = range.splitEvenly(255);
         assertThat(ranges).hasSize(255);
 
-        for(int i = 0; i < ranges.size(); i++) {
+        for (int i = 0; i < ranges.size(); i++) {
             TokenRange tr = ranges.get(i);
-            if(i < 10) {
-                assertThat(tr).isEqualTo(tokenRange(i, i+1));
+            if (i < 10) {
+                assertThat(tr).isEqualTo(tokenRange(i, i + 1));
             } else {
                 assertThat(tr.isEmpty());
             }